requirements.txt
customtkinter==5.2.1
Pillow==10.0.0
python-dateutil==2.8.2
pyrebase4==4.7.0
python-dotenv==1.0.0
